requirements.txt

google.cloud.storage==1.31.0
requests==2.27.1
requests-oauthlib==1.3.1
yfinance==0.1.70
seaborn==0.11.2
squarify==0.4.3
matplotlib==3.5.1
pandas-datareader==0.10.0
cloudstorage==0.11.0
import gcsfs